Fix toJSON for LazyMessage array elements and fields
https://github.com/foxglove/rosmsg-serialization/pull/54 regressed on toJSON
functionality for fields and array elements of a lazy message which are
themselves lazy messages. While the top level LazyMessage toJSON correctly
serialized the message to a javascript object, accessing a custom field or an
array element in the lazy message and calling toJSON on it would incorrectly produce
the top level message rather than the individual field.

This change fixes the toJSON behavior so each instance of a LazyMessage properly
serializes itself to a javascript object.