Fix encoding of non-ascii strings (#361)
### Public-Facing Changes

Fixed a bug where non-ASCII strings would be truncated when `MessageWriter` encoded them as UTF-8.

### Description

Fixes #360

`str.length` is the length of the string in UTF-16. However, we use TextEncoder to encode strings as UTF-8. We could compute the length using TextEncoder itself, but this proves to be much slower than a manual calculation: https://jsbench.me/nzlrkwmeiq/1